Rename GrpcService to GrpcClient
