+++ src/Worker.ts
@@ -4,7 +4,7 @@
 import * as parseArgs from 'minimist';
 import { AzureFunctionsRpcMessages as rpc } from '../azure-functions-language-worker-protobuf/src/rpc';
 import { FunctionLoader } from './FunctionLoader';
-import { CreateGrpcEventStream } from './GrpcService';
+import { CreateGrpcEventStream } from './GrpcClient';
 import { InternalException } from './utils/InternalException';
 import { systemError, systemLog } from './utils/Logger';
 import { WorkerChannel } from './WorkerChannel';
